代码重构，天体样式封装
