+++ packages/celestial-demo/src/styles/celestial-bodies/index.ts
@@ -1,0 +1,9 @@
+// 集中导出所有天体创建函数
+export { createPlanet } from './planet';
+export { createStar } from './star';
+export { createBlackHole } from './black-hole';
+export { createMoon } from './moon';
+export { createComet } from './comet';
+export { createNeutronStar } from './neutron-star';
+export { createNebula } from './nebula';
+export { createGalaxy } from './galaxy';
